python/tracing/llamaindex-openinference/requirements.txt
llama-index-core
llama-index-llms-openai
openinference-instrumentation-llama-index>=4.0
respan-tracing
respan-instrumentation-openinference
python-dotenv
